requirements-ci.txt
google-cloud-secret-manager==0.2.0
semver==2.8.1
